+++ R/distsfactory-package.R
@@ -19,10 +19,10 @@
 #' @name distsfactory-package
 #' @aliases distsfactory
 #' @importFrom stats dbeta dbinom dcauchy dchisq dexp df dgamma dgeom dlnorm
-#'   dlogis dnbinom dnorm dpois dt dunif dweibull integrate pbeta pbinom
-#'   pcauchy pchisq pexp pf pgamma pgeom plnorm plogis pnbinom pnorm ppois pt
-#'   punif pweibull qbeta qbinom qcauchy qchisq qexp qf qgamma qgeom qlnorm
-#'   qlogis qnbinom qnorm qpois qt qunif qweibull rbeta rbinom rcauchy rchisq
-#'   rexp rf rgamma rgeom rlnorm rlogis rnbinom rnorm rpois rt runif rweibull
-#'   uniroot
+#'   dlogis dnbinom dnorm dpois dt dunif dweibull integrate median pbeta
+#'   pbinom pcauchy pchisq pexp pf pgamma pgeom plnorm plogis pnbinom pnorm
+#'   ppois pt punif pweibull qbeta qbinom qcauchy qchisq qexp qf qgamma qgeom
+#'   qlnorm qlogis qnbinom qnorm qpois qt quantile qunif qweibull rbeta rbinom
+#'   rcauchy rchisq rexp rf rgamma rgeom rlnorm rlogis rnbinom rnorm rpois rt
+#'   runif rweibull uniroot
 "_PACKAGE"